Transfer money between accounts
An API endpoint is created for the flow of a money transfer between
bank accounts. Fail cases are considered, but without the proper error
handling yet.

`versionist` gem is used for API versioning, and `jbuilder` is used to
create the JSON response, which follows JSON:API specification.

See https://jsonapi.org/format/#document-resource-objects